// windows-admin-center-developer-tools/inlineSrc/generated/strings.ts
/* tslint:disable */
/**
 * @file Source code generated by gulp-resjson.
 * @version 1.0
 */
export interface Strings {
    HelloWorld: {
        title: string;
        service: string;
        notFound: string;
        longRunning: string;
        error2Running: string;
        error7Running: string;
        exception2Running: string;
        exception7Running: string;
        clientNotification: string;
        showCode: string;
        hideCode: string;
        about: string;
        aboutDescription: string;
        Manifest: {
            displayName: string;
            description: string;
            connectionsListHeader: string;
            overviewName: string;
            overviewDescription: string;
            keywordExample: string;
            devDescription: string;
            devDisplayName: string;
            solutionName: string;
            solutionDescription: string;
        };
        nav: {
            controls: string;
            alerts: string;
            styles: string;
        };
        powershell: {
            title: string;
        };
        notifications: {
            title: string;
        };
        dll: {
            title: string;
        };
        cim: {
            title: string;
            viewCode: string;
            hideCode: string;
        };
        controls: {
            title: string;
            subtitle: string;
            showTreeCode: string;
            showDataTableCode: string;
        };
    };
};